pg: pi: brcm: attempt to correct es2 window placement
On some screen setups piworld was showing its 3D display either badly
aligned or not at all.

Fix a bug in which width and height are unsigned int's, but the result
for xy position is a signed int by casting width and height to an
unsigned int for xy calculations.

+++ deps/pg/pg_pi.c
@@ -189,8 +189,8 @@ void pi_set_window_geometry(int *x, int *y, uint32_t *w, uint32_t *h)
 
     // Account for screen offset
     screen = DefaultScreenOfDisplay(pi_state->x11_display);
-    *x += (pi_state->screen_w - WidthOfScreen(screen)) / 2;
-    *y += (pi_state->screen_h - HeightOfScreen(screen)) / 2;
+    *x += ((int)pi_state->screen_w - WidthOfScreen(screen)) / 2;
+    *y += ((int)pi_state->screen_h - HeightOfScreen(screen)) / 2;
 
     int reduction = get_reduction_factor(*w, *h);
